example/live/requirements.txt
sanic
python-socketio
